size_t loop var in drawminimap()

+++ src/engine/render/rendergl.cpp
@@ -1419,7 +1419,7 @@ void drawminimap(int yaw, int pitch, vec loc, const cubeworld& world, int scalef
     }
     ivec bbmin(rootworld.mapsize(), rootworld.mapsize(), rootworld.mapsize()),
          bbmax(0, 0, 0);
-    for(uint i = 0; i < valist.size(); i++)
+    for(size_t i = 0; i < valist.size(); i++)
     {
         const vtxarray *va = valist[i];
         for(int k = 0; k < 3; ++k)
